mv start up code into main
