+++ northstar/__init__.py
@@ -1,55 +1,3 @@
 """
 North Star Application
 """
-# North Star ---  A lookup service for forged fed ecosystem
-# Copyright © 2021 Aravinth Manivannan <realaravinth@batsense.net>
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-import os
-import sqlite3
-
-from flask import Flask
-
-from . import db
-from .api import V1_bp
-
-
-def create_app(test_config=None):
-    # create and configure the app
-    app = Flask(__name__, instance_relative_config=True)
-    app.config.from_mapping(
-        DATABASE=os.path.join(app.instance_path, "northstar.db"),
-    )
-
-    db.init_app(app)
-
-    if test_config is None:
-        app.config.from_pyfile("config.py", silent=True)
-    else:
-        app.config.from_mapping(test_config)
-
-    try:
-        os.makedirs(app.instance_path)
-    except OSError:
-        pass
-
-    @app.after_request
-    def flock_google(response):
-        response.headers["Permissions-Policy"] = "interest-cohort=()"
-        return response
-
-    app.register_blueprint(V1_bp)
-    return app
-
-
-app = create_app()
