+++ northstar/__main__.py
@@ -1,5 +1,7 @@
-""" Test errors helper class"""
-# North Star ---  A lookup service for forged fed ecosystem
+"""
+Run ForgeFed Interface flask application
+"""
+# Bridges software forges to create a distributed software development environment
 # Copyright © 2021 Aravinth Manivannan <realaravinth@batsense.net>
 #
 # This program is free software: you can redistribute it and/or modify
@@ -11,21 +13,12 @@
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU Affero General Public License for more details.
+#
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 from northstar.app import create_app
 
-from northstar.api.v1.errors import Error
-from northstar.api.v1.interface import F_D_EMPTY_FORGE_LIST, F_D_INTERFACE_UNREACHABLE
-
-
-def test_errors(client):
-    """Test interface registration handler"""
-
-    def verify_status(e: Error, status: int):
-        assert e.status() == status
-        resp = e.get_error_resp()
-        assert resp.status.find(str(status)) is not -1
-
-    verify_status(F_D_EMPTY_FORGE_LIST, 400)
-    verify_status(F_D_INTERFACE_UNREACHABLE, 503)
+if __name__ == "__main__":
+    app = create_app()
+    app.run(threaded=True, port=7000)
