+++ tests/test_default_headers.py
@@ -13,7 +13,7 @@
 # GNU Affero General Public License for more details.
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from northstar import create_app
+from northstar.app import create_app
 
 
 def test_flock_google(client):
